% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRextractors.R
\name{expectedFDR_trueFDR_power}
\alias{expectedFDR_trueFDR_power}
\title{compute expeted true FDR and power}
\usage{
expectedFDR_trueFDR_power(pvalue, outlier)
}
\arguments{
\item{pvalue}{a vector of pvalue for each snips}

\item{outlier}{a vector of outlier snps index}
}
\description{
assumig we know m0 the number not outlier. We use
P(H0|positif) = P(positif | H0) * P(H0) / P(positif)
}
