% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-fromTrueData.R
\name{ExpRsampler_fromTrueData}
\alias{ExpRsampler_fromTrueData}
\title{see my notebook (13/06/2017), sample a generative model from true dataset.}
\usage{
ExpRsampler_fromTrueData(Y, K, prop.outlier, cs = NULL, rho.c = 1,
  rho.B = 1, n = NULL, p = NULL)
}
\description{
see my notebook (13/06/2017), sample a generative model from true dataset.
}
\author{
cayek
}
