% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-generativeData.R
\name{ExpRsampler_generativeData}
\alias{ExpRsampler_generativeData}
\title{Generate data from the generative model of lfmm.}
\usage{
ExpRsampler_generativeData(n, p, K, outlier.prop, cs, sigma = 0.2, B.sd = 1,
  B.mean = 0, U.sd = 1, V.sd = 1)
}
\description{
Generate data from the generative model of lfmm.
}
\author{
cayek
}
