% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-generativeData.R
\name{ExpRsampler_generativeData_BVcorrelated}
\alias{ExpRsampler_generativeData_BVcorrelated}
\title{Generate data from the generative model of lfmm.}
\usage{
ExpRsampler_generativeData_BVcorrelated(n, p, K, outlier.prop, csXU, csBV,
  sigma = 0.2)
}
\description{
Generate data from the generative model of lfmm.
}
\author{
cayek
}
