% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-trueData.R
\name{ExpRsampler_trueData}
\alias{ExpRsampler_trueData}
\title{Sample true data set from file.}
\usage{
ExpRsampler_trueData(Y, X, outlier)
}
\description{
Sample true data set from file.
}
\author{
cayek
}
