% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-fromTrueData.R
\name{load_sampler_fromTrueData}
\alias{load_sampler_fromTrueData}
\title{Load into an environment true data}
\usage{
load_sampler_fromTrueData(s)
}
\description{
Load into an environment true data
}
