% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpRsampler-trueData.R
\name{ExpRsampler_trueData}
\alias{ExpRsampler_trueData}
\title{.. content for \description{} (no empty lines) ..}
\usage{
ExpRsampler_trueData(Y, X, outlier)
}
\description{
.. content for \details{} ..
}
\author{
cayek
}
